Refractor each file to be more object oriented
