+++ part-3/book.rb
@@ -1,39 +1,24 @@
+require_relative 'items'
+
 class Book
   attr_reader :length, :title, :author, :format, :language
 
+  include RentalOption
+
   def initialize(args = {})
-    @length = args[:length]
-    @title  = args[:title]
-    @author = args[:author]
-    @format = args[:format]
-    @language = args[:language]
+    parameter_group(args)
   end
 
-  def available?
-    !checked_out?
+  private
+
+  attr_writer :length, :title, :author, :format, :language
+
+  def parameter_group(args)
+    self.length = args[:length]
+    self.title  = args[:title]
+    self.author = args[:author]
+    self.format = args[:format]
+    self.language = args[:language]
   end
 
-  def checkout
-    @checked_out = true
-  end
-
-  def checked_out?
-    @checked_out
-  end
-
-  def return
-    @checked_out = false
-  end
-
-  def damaged?
-    @damaged
-  end
-
-  def record_damage
-    @damaged = true
-  end
-
-  def repair
-    @damaged = false
-  end
 end
