+++ part-3/c_d.rb
@@ -1,4 +1,9 @@
+require_relative 'items'
+
 class CD
+
+  include RentalOption 
+
   attr_reader :runtime, :title, :artist, :label
 
   def initialize(args ={})
@@ -8,31 +13,4 @@ class CD
     @label   = args[:label]
   end
 
-  def available?
-    !checked_out?
-  end
-
-  def checkout
-    @checked_out = true
-  end
-
-  def checked_out?
-    @checked_out
-  end
-
-  def return
-    @checked_out = false
-  end
-
-  def damaged?
-    @damaged
-  end
-
-  def record_damage
-    @damaged = true
-  end
-
-  def repair
-    @damaged = false
-  end
 end
