+++ part-3/movie.rb
@@ -1,4 +1,8 @@
+require_relative 'items'
+
 class Movie
+  include RentalOption
+
   attr_reader :runtime, :title, :format
 
   def initialize(args ={})
@@ -7,31 +11,4 @@ class Movie
     @format  = args[:format]
   end
 
-  def available?
-    !checked_out?
-  end
-
-  def checkout
-    @checked_out = true
-  end
-
-  def checked_out?
-    @checked_out
-  end
-
-  def return
-    @checked_out = false
-  end
-
-  def damaged?
-    @damaged
-  end
-
-  def record_damage
-    @damaged = true
-  end
-
-  def repair
-    @damaged = false
-  end
 end
