+++ part-3/video_game.rb
@@ -1,4 +1,9 @@
+require_relative 'items'
+
 class VideoGame
+
+  include RentalOption
+
   attr_reader :platform, :title, :developer, :genre
 
   def initialize(args = {})
@@ -8,31 +13,4 @@ class VideoGame
     @genre = args[:genre]
   end
 
-  def available?
-    !checked_out?
-  end
-
-  def checkout
-    @checked_out = true
-  end
-
-  def checked_out?
-    @checked_out
-  end
-
-  def return
-    @checked_out = false
-  end
-
-  def damaged?
-    @damaged
-  end
-
-  def record_damage
-    @damaged = true
-  end
-
-  def repair
-    @damaged = false
-  end
 end
